Finish implementation of TerraModel, without interpolation
This commit completes the start of a TerraModel class with some
useful functionality.  It lacks docs.

TerraModel is a class which holds arbitrarily-located lateral points
which are repeated at several levels, giving a 2D (or 3D for vector
fields or multi-component fields) array for each field held on that
set of points.  It is agnostic to where the radial layers are and
at present where the points are.  Hence it can support points in a small
region of the sphere.

The important contributions are:
- A `read_netcdf` method to read a `TerraModel` from several NetCDF
  files.
- Support for composition histograms, where some number of compositional
  components are used to represent a mechanical mixture of different
  rock types at each point.
- Ability to access each field via a stable API.
- Likewise, a stable API for getting information about the coordinates
  of a model.
- Ability to correctly create new fields of the appropriate dimensions.
- Restriction of which fields you can create, and automatic
  determination of what dimensions any field should have.

It currently lacks:
- Any support for the TERRA mesh in particular.
- Ability to create meshes for appropriate TERRA subdivision levels.
- Support for interpolation of field values between points.

Notable current limitations:
- It's quite slow to read in all the NetCDF files.