terra_model: Implement interpolation for `evaluate`
Implement `TerraModel.evaluate`.  This returns the value of a field
at an arbitrary point in the model, using either nearest-neighbour
lookup or interpolation.  It assumes for the latter that the lateral
points are arranged such that the three nearest points form a triangle
around the point of interest.  No check is made for this.

As a side effect, also expose `TerraModel.nearest_layer` to give
the layer index and radius of the closes layer to a radius of interest.

Also add examples to show how this interpolation (or nearest-neighbour
evaluation) works with plots.