lookup_tables: Fix bug in _harmonic_mean
Previously, _harmonic_mean would give the wrong answer where lists of
multi-dimensional numpy arrays were passed in.  This was because of the
call to `np.sum(fractions)`, which simply sums the entire array passed
in.

Fix this by keeping track of the total fractions (weightings) in a
separate array (or scalar).  Use of `np.shape` on input arguments allows
for either scalars, arrays or lists of arrays to be passed in.