plot.layergrid: Only allow coastlines via cartopy, not not at all
Rename the `cartopy_coastlines` keyword argument to just `coastlines`,
and if it is `False`, don't plot coastlines at all.

This change was made because it transpires that if you have not
installed cartopy either via Conda or building from source, you hit a
segfault when calling `ax.coastlines()`.  The same segfault which
causes problems with cartopy's own `coastlines()` function also hits you
if you try to plot any lines on a `GeoAxis`, so it is impossible to even
plot some other custom coastlines.  Therefore the only workaround is to
not plot any lines on the plot, and call `coastlines=False` on affected
machines.