Make cartopy an optional dependency
This means that users must install cartopy separately if they want to
make maps which use cartopy, via `pip` or otherwise.

Update the test script to install cartopy unconditionally.  A future
commit will add a test for the correct behaviour (an `ImportError`)
when trying to call a function which relies on cartopy.