terra_model: Rename `get_1d_mean` to `mean_1d_profile`
The old name `get_1d_mean` implied via the `get_` that some part of the
model was being returned that could be used to update the `TerraMode`
from which it came.  However, at odds with the convention it does not
do that, so rename it to a 'noun-type' name which returns the thing it
is named after.