TerraModel.plot_section: Fix error with small radius ranges
For models where the difference between the largest and smallest
radii is smaller than the default radial spacing in
`terra_model.TerraModel.plot_section`, previously an error was
thrown when trying to plot because only one radial point was
evaluated in the grid.  Fix this by ensuring at least two radial
points are in the grid.