Move example table data files to figshare
This introduces a new submodule that is designed to manage
caching of example data. The module (example_data) has functions
for each file we need to be able to access, on first use the file
will be downloaded, for other uses the file is read out of a local
cache.

This is used to manage the files: example_bas_table.dat,
example_hzb_table.dat and example_lhz_table.dat used in the
lookup table example. We need to modify test.sh to avoid failing
when download information is reported.