Docs: Update `TerraModel.evalute` example for example model
Prviously, the `examples/example_evaluate_points.py` script
used a simple made-up model to show how to evaluate points.  Now
that we use a real example model, update the script to use a more
interesting part of the model to show how to use `.evaluate()`.
The existing coordinates used did not show any interesting
variation in the new model file.