TerraModel.radial_profile: Rename `1d_profile` methods
Previously, we had a `get_` method which returned the radial or 1D
profile of a model at a certain point, but `get_` methods are by
convention those which return the data inside a model, possibly
for editing.  In updating this name, a simple change to
`1d_profile` was not possible under Python's naming rule so another
name was needed.  This commit changes `TerraModel.get_1d_profile` to
`TerraModel.radial_profile`, and for consistency also changes
`TerraModel.mean_1d_profile` to `TerraModel.mean_radial_profile`.

As an alternative I considered `profile_1d`, but that didn't feel
right.  A '1D profile' is really tautologous, since a profile is by
definition 1D.  Hence adding information about in which direction
the profile was running seemed most useful.  Finally therefore we
settle on `TerraModel.radial_profile` as the best alternative.