Retry example data downloads
The example data 'distributed' with terratools is actually stored
on Figshare and downloaded (and cached) on demand.  This can cause
problems on CI because multiple runners all try and download the
same file many times simultaneously, and Figure reasonably starts
to reject the requests.  As a workaround, by default retry the
download a few times before giving up.

May address the following issue:
https://github.com/mantle-convection-constrained/terratools/issues/185