Turn on data caching for tests
Sometimes out tests fail because pooch fails to download
test data to the local runner from Figshare. This seems to
be a load issue. This commit makes use of the github actions
caching tools to create and maintain a local copy of the test
data. We try to restore from this cache before running the tests
and then write back to the cache afterwards (so we should add
any extra data we may have downloaded). The cache should work
between OSs and python versions but we don't explicity enable
this. When we write to the cache we add the job id. When we read
we should get the closest copy.

This is a first pass and could probably be improved to include
hashes of the data files, but this ought to mostly get the job
done and our data isn't that big!