Update the pyproject.toml to use the latest conventions
`pyrpoject.toml` determines requirements and has got to the point where
standard information such as authorship has been standardised.
Therefore include this information in an updated `pyproject.toml` which
can then be used by tools such as `poetry` and `uv` to build releases
and manage dependencies.

This requires some extra instruction when installing the package
locally for development when using `pip`.  Add this information to
`README.md` and use the correct incantation in our test and docs
scripts.

At the same time, and to determine whether testing is working, update a
couple of things:
- Remove the unused `cgit` import in tests which has been removed in
  recent Python versions as a stdlib.
- Fix a deprecation in `sklearn.clustering.HDBSCAN` which was causing
  test errors.